refactor(reco-core): promote StitchExecutor (rename, crate-internal until wired)
Step 5 of the stitch-unification roadmap: StitchBackend -> StitchExecutor,
Cpu/GpuStitchBackend -> Cpu/GpuExecutor (backend.rs -> executor.rs), the
stitch() signature untouched - it reshapes when the Projection trait
starts dispatching geometry. The executors and the L-shape gather
internals drop to crate visibility: their first real consumer is the
CPU-executor step, and publishing a callerless surface that still
reshapes would freeze it prematurely. The agreement oracle and no-black
property test keep the whole chain exercised; the module-level
dead_code allow marks the scaffolding as deliberate.
